Move bookmark component to its own directory
